requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
streamlit>=1.32.0
requests>=2.31.0
pydantic>=2.6.0
PyYAML>=6.0.1
SQLAlchemy>=2.0.27
pymysql>=1.1.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-chroma>=0.1.0
langgraph>=0.2.0
dashscope>=1.15.0
